#507 move maine page to pages
